Make deposits set the contentprovider correctly.
The content_provider_id property was being ignored by doctrine. When a
class defines a field for a foreign key and a manyToOne/oneToMany field
doctrine ignores the foreign key field and uses the manyToOne field
instead. But the code was only setting the foreign key field, so the
content provider was never set for the deposit.

Now on to testing that the deposit's fields are set correctly. Must
think about it.